[Feat] #20 MainView 비디오 기능 코드 분리 (모델, 뷰모델은 임시로 만든 구조입니다.)
